+++ apps/web/webpack.config.js
@@ -10,6 +10,7 @@ module.exports = (_, { mode }) => ({
   },
   resolve: {
     extensions: ['.ts', '.tsx', '.js', '.jsx'],
+    alias: { '@': path.resolve(__dirname, 'src') },
   },
   module: {
     rules: [